backend/requirements.txt
fastapi==0.109.2
gunicorn
uvicorn[standard]

python-dotenv

pymongo
dnspython

python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

requests

pydantic<2
email-validator
python-multipart

Pillow

azure-storage-blob
python-dateutil

stripe
